Sends all presence messages to _presence_handlers
Currently any presence message that does not have
an ‘x’ tag does not get send to the
presence_handler. This means error messages such
as

<presence xmlns='jabber:client' type='error' to=‘xx’ from=‘xx/test'>
<error type='cancel'>
  <not-allowed xmlns='urn:ietf:params:xml:ns:xmpp-stanzas'/>
</error>
</presence>

does not get passed to the handler. This results in
errors like this not being handled and client code
not knowing about it and instead assuming
everything is fine.

This commit removes the custom handling and passes
all presence messages to the handlers attached,
which is how the code has always worked in the case
of the message_handlers.

This way error handling can be implemented in the
presence_handlers, as how they are being done in
the message_handlers.